// src/patients/entities/education-level.enum.ts
export enum EducationLevel {
  INITIAL = 'INITIAL',
  PRIMARY_INCOMPLETE = 'PRIMARY_INCOMPLETE',
  PRIMARY = 'PRIMARY',
  SECONDARY_INCOMPLETE = 'SECONDARY_INCOMPLETE',
  SECONDARY = 'SECONDARY',
  TECHNICAL = 'TECHNICAL',
  TECHNICAL_INCOMPLETE = 'TECHNICAL_INCOMPLETE',
  HIGHER = 'HIGHER',
  HIGHER_INCOMPLETE = 'HIGHER_INCOMPLETE',
  NONE = 'NONE',
}